BSH-73 add image handling for players

+++ routers/players.py
@@ -23,6 +23,7 @@ router = APIRouter()
 auth = AuthHandler()
 configure_cloudinary()
 
+
 # upload file
 async def handle_image_upload(image: UploadFile, playerId) -> str:
   if image:
@@ -34,18 +35,26 @@ async def handle_image_upload(image: UploadFile, playerId) -> str:
         resource_type="image",
         format='jpg',  # Save as JPEG
         transformation=[{
-            'effect': 'background_removal'
-        }, {
-            'width': 300, 
-            'height': 300, 
+            'width': 300,
+            'height': 300,
             'crop': 'thumb',
             'gravity': 'face'
-        }]
-    )
+        }])
     print(f"Player image uploaded to Cloudinary: {result['public_id']}")
     return result["secure_url"]
   raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
                       detail="No image uploaded.")
+
+async def delete_from_cloudinary(image_url: str):
+  if image_url:
+    try:
+      public_id = image_url.rsplit('/', 1)[-1].split('.')[0]
+      result = cloudinary.uploader.destroy(f"players/{public_id}")
+      print("Document deleted from Cloudinary:", f"players/{public_id}")
+      print("Result:", result)
+      return result
+    except Exception as e:
+      raise HTTPException(status_code=500, detail=str(e))
 
 
 # Helper function to search players
@@ -771,18 +780,17 @@ async def create_player(
   # Generate a new ID for the player
   player_id = str(ObjectId())
 
-  player = PlayerBase(
-                     firstName=firstName,
-                     lastName=lastName,
-                     birthdate=birthdate,
-                     displayFirstName=displayFirstName,
-                     displayLastName=displayLastName,
-                     nationality=nationality,
-                     position=position,
-                     assignedTeams=assigned_teams_dict,
-                     fullFaceReq=fullFaceReq,
-                     source=SourceEnum[source],
-                     legacyId=legacyId)
+  player = PlayerBase(firstName=firstName,
+                      lastName=lastName,
+                      birthdate=birthdate,
+                      displayFirstName=displayFirstName,
+                      displayLastName=displayLastName,
+                      nationality=nationality,
+                      position=position,
+                      assignedTeams=assigned_teams_dict,
+                      fullFaceReq=fullFaceReq,
+                      source=SourceEnum[source],
+                      legacyId=legacyId)
   player = my_jsonable_encoder(player)
   player['create_date'] = datetime.now().replace(microsecond=0)
   player['_id'] = player_id
@@ -795,13 +803,13 @@ async def create_player(
     created_player = await mongodb["players"].find_one({"_id": player_id})
     if created_player:
       return JSONResponse(status_code=status.HTTP_201_CREATED,
-                         content=jsonable_encoder(PlayerDB(**created_player)))
+                          content=jsonable_encoder(PlayerDB(**created_player)))
     else:
       raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                         detail="Failed to create player.")
+                          detail="Failed to create player.")
   except Exception as e:
     raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                       detail=str(e))
+                        detail=str(e))
 
 
 # UPDATE PLAYER
@@ -892,10 +900,13 @@ async def delete_player(
   mongodb = request.app.state.mongodb
   if token_payload.roles not in [["ADMIN"]]:
     raise HTTPException(status_code=403, detail="Not authorized")
-
+  existing_player = await mongodb["players"].find_one({"_id": id})
+  if not existing_player:
+    raise HTTPException(status_code=404,
+                        detail=f"Player with id {id} not found")
   delete_result = await mongodb["players"].delete_one({"_id": id})
   if delete_result.deleted_count == 1:
+    await delete_from_cloudinary(existing_player['imageUrl'])
     return Response(status_code=status.HTTP_204_NO_CONTENT)
-  else:
-    raise HTTPException(status_code=404,
+  raise HTTPException(status_code=404,
                         detail=f"Player with ID {id} not found.")
